refactor: 우승자 로직을 RacingCarGame으로 이동
