+++ src/test/kotlin/racingcar/model/CarInitializerTest.kt
@@ -1,0 +1,12 @@
+package racingcar.model
+
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Test
+
+class CarInitializerTest {
+    @Test
+    fun `Car 생성`() {
+        val cars = CarInitializer.createCars(listOf("pobi", "crong"))
+        assertThat(cars.size).isEqualTo(2)
+    }
+}
