+++ src/test/kotlin/racingcar/model/RacingCarGameTest.kt
@@ -21,7 +21,23 @@ class RacingCarGameTest {
     }
 
     @Test
-    fun `RacingHistory의 round는 0부터 총 횟수 - 1 까지 생성된다`() {
+    fun `차들 중 가장 멀리 위치한 차가 우승자가 된다`() {
+        // given
+        val firstCar = Car("car1", 2)
+        val secondCar = Car("car2", 4)
+        val racingHistories = listOf(RacingHistory(0, listOf(firstCar, secondCar)))
+
+        // when
+        val winner = RacingCarGame.getWinners(racingHistories)
+
+        // then
+        assertThat(winner.size).isEqualTo(1)
+        assertThat(winner.first().name).isEqualTo("car2")
+        assertThat(winner.first().position).isEqualTo(4)
+    }
+
+    @Test
+    fun `5번을 전진시도할 때 게임이 시작되면 0부터 4까지 5개의 게임 기록이 존재해야한다`() {
         // given
         val cars = arrayListOf(Car("myCar"))
         val strategy = FakeMovingStrategy(true)
